ProjectFastAPI/requirements.txt
uvicorn
fastapi
sqlalchemy
python-dotenv
psycopg2
alembic
passlib
python-jose
python-multipart
flake8